Slideshow now exists with crossfading (!!) thanks to millstream's fantastically amazing "Crossfade Anything" script. Source files are nice and tidy, almost ripe for localization.
Slides are defined in slide-base.html. Just add a div inside "container" and gaze in awe. A slide is now contained inside a small (1kb) file, with slide-template.html as an example. Fill in the blanks and you're done.

TODO: main_shadow_top and main_shadow_bottom should not be needed in the template. Find a way to implement these in CSS.