Enlarging menu icons to 24px, as is used in the real main menu. This way they look clear and pretty.
The down side: Anything referencing a menu is going to look bad if it has a line right below it, because it really sticks out. So try to keep these lines to themselves.

Simplified text in translations.html further, removing mention of needing to log in again since those directions are right there in the tool.