Holy talking cows, Batman, it works! In a horrible, proof of concept kind of way.
Added Makefile, which will automatically create translated versions of the slideshow out of files in the po directory. This is my first Makefile from scratch, so please feel free to educate me on how broken it is.

Renamed License and Readme files to COPYING and README, respectively, to better fit with convention.