All slideshows: Moved hash parameter parsing to base.js (from slideshow.js) Added signals slideshow-started, slide-opening, slide-opened, slide-closing and slide-closed Added data-translate attribute, which is understood by po4a as translatable
Ubuntu:
Moved some more Twitter presentation stuff to a TwitterStream object.
Twitter stream is loaded as soon as slideshow starts, but only shows tweets when its respective slide is visible. Otherwise, it is paused.
Twitter stream determines if there is a live connection as soon as slideshow starts and hides or shows itself depending on the result.
"I'm installing #Ubuntu" status text can now be translated.